refactor: error 타입 생성 및 decode 에러처리 수정 및
