+++ Expo1900/Expo1900/DecodeError/Error.swift
@@ -1,0 +1,13 @@
+//
+//  Error.swift
+//  Expo1900
+//
+//  Created by Seoyeon Hong on 2023/02/28.
+//
+
+import Foundation
+
+enum DecodeError: Error {
+    case expositionDecodeError
+    case expoItemDecodeError
+}
